index_v2.js
before: 
unable to spam the drum sound during keypress because the program still waiting for the sound to finish and it may feel not 
like an actual drum. 

after: 
during keypress,  resets the playtime of the sound in order to spam the sound to simulate the feel of a real drum.